% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file_name_materials_experiment3.R
\name{get_file_name_materials_experiment3}
\alias{get_file_name_materials_experiment3}
\title{Get E3 screenshot file names}
\usage{
get_file_name_materials_experiment3()
}
\value{

}
\description{
Get E3 screenshot file names
}
\author{
Shir Dekel
}
