% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/get_project_description_similarity_experiment3.R
\name{get_project_description_similarity_experiment3}
\alias{get_project_description_similarity_experiment3}
\title{Get long project descriptions for similarity conditions}
\usage{
get_project_description_similarity_experiment3(
  gambles,
  project_detail_experiment3,
  project_components_experiment3,
  label_similarity
)
}
\arguments{
\item{gambles}{}

\item{project_components_experiment3}{}
}
\value{

}
\description{
Get long project descriptions for similarity conditions
}
\author{
Shir Dekel
}
