% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_projects_experiment3.R
\name{get_projects_experiment3}
\alias{get_projects_experiment3}
\title{Get long projects}
\usage{
get_projects_experiment3(gambles)
}
\arguments{
\item{gambles}{}
}
\value{

}
\description{
Get long projects
}
\author{
Shir Dekel
}
