% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_trials_experiment3.R
\name{get_trials_experiment3}
\alias{get_trials_experiment3}
\title{Get long projects distribution absent trial}
\usage{
get_trials_experiment3(projects_experiment3, randomize_order)
}
\arguments{
\item{projects_experiment3}{}

\item{randomize_order}{}
}
\value{

}
\description{
Get long projects distribution absent trial
}
\author{
Shir Dekel
}
