% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregation3.R
\docType{package}
\name{aggregation3}
\alias{aggregation3}
\alias{aggregation3-package}
\title{Aggregation 3 experiment}
\description{
Data, materials, and experiment code for the Aggregation 3 experiment.
}
\details{
aggregation3
A data package for the Aggregation 3 experiment.

Use \code{data(package='aggregation3')$results[, 3]} to see a list of available data sets in this data package
and/or DataPackageR::load_all
_datasets() to load them.
}
\seealso{
\link{data}
}
